showreview/user/requirements.txt
psycopg2-binary
django
djangorestframework
pytest
pytest-django
PyJWT
pika